+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/gwo_panel.js
@@ -115,7 +115,7 @@ function gwoWarInfoPanel(gwoSettings) {
       ["coui://ui/mods/com.pa.quitch.gwaioverhaul/gw_play/commander_colour.js"],
       function (gwoColour) {
         /* War Information */
-        model.gwoVersion = ko.observable("5.91.0");
+        model.gwoVersion = ko.observable("5.92.0");
 
         /* Co-op Information */
         const coopText = function (setting) {